tests: partner detail view
