+++ templates/partnerships/includes/contact.html
@@ -2,8 +2,8 @@
 <p>
   {% trans "contact_mobility_in" %}<br>
   <strong>{{ contact }}</strong> {{ contact.function }}{% if contact.function and contact.society %}, {% endif %}{{ contact.society }}<br>
-  {% if contact.phone %}<i class="fa fa-fw fa-phone"></i> {{ contact.phone }}{% endif %}
-  {% if contact.mobile_phone %}<i class="fa fa-fw fa-mobile"></i> {{ contact.mobile_phone }}{% endif %}
-  {% if contact.fax %}<i class="fa fa-fw fa-fax"></i> {{ contact.fax }}{% endif %}<br>
+  {% if contact.phone %}<i class="fa fa-fw fa-phone" title="{% trans "phone" %}"></i> {{ contact.phone }}{% endif %}
+  {% if contact.mobile_phone %}<i class="fa fa-fw fa-mobile" title="{% trans "mobile_phone" %}"></i> {{ contact.mobile_phone }}{% endif %}
+  {% if contact.fax %}<i class="fa fa-fw fa-fax" title="{% trans "fax" %}"></i> {{ contact.fax }}{% endif %}<br>
   {% if contact.email %}{{ contact.email|urlize }}{% endif %}
 </p>
